requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1639166893812/work
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.14.6
cryptography==36.0.1
cssselect2==0.4.1
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
Pillow==8.3.1
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.20
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tzdata==2021.5
weasyprint==53.0
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.5
XlsxWriter==3.0.2
